Some changes about my frustrations learning this
You'd think that it'd be fairly simple - but you don't even want to know how difficult it was for me to get here. But finally I found this wonderful little page which explained it all out to me. So here we are and there we are and the future looks limitless.